Fix rectangle corners
It turns out that zxing returns `ResultPoints` in a weird order, so we
have to pick those apart.

We're now able to clearly show where the code is detected, but there's
still a glaring issue: it only detects codes that are more or less
centered in the frame.  I haven't figured out why yet, but we'll need to
fix that before we can reliably test with this.