Compute detection candidates from transitions
Our first pass gets us the locations at which greyscale value changes
abruptly.  In this second pass, we now filter and score those
transitions, looking for consecutive runs of three transitions along the
same ray with high strength and roughly equal spacing.

To achieve this, we use a heuristic that adds the transition strengths,
then subtracts a measure of how irregularly spaced the transitions are.

From here, we'll need to compute our best guess at the two shoulder
locations.  One thought here is a weighted average of candidates on each
side; we'll probably try that, assess it for accuracy, then tweak as
necessary!